+++ utils/fetch-version.js
@@ -4,7 +4,7 @@ import axios from 'axios';
  * @returns {Promise<string>}
  */
 const fetchVersion = async () => {
-  const { data } = await axios.get('https://raw.githubusercontent.com/memochou1993/gpt-ai-assistant/main/package.json');
+  const { data } = await axios.get('https://raw.githubusercontent.com/bigcan/gpt-ai-assistant/main/package.json');
   return data.version;
 };